knowledge-ingesters/document_extractor/requirements_multi_format.txt
# Multi-Format Knowledge Extraction Pipeline Dependencies
# Install with: pip install -r requirements_multi_format.txt

# Core Dependencies (already in project)
anthropic>=0.34.0
weaviate-client>=4.0.0
python-dotenv>=1.0.0

# PDF Processing
PyMuPDF>=1.26.4  # fitz - PDF page conversion
Pillow>=11.3.0   # Image processing

# DOCX Processing
python-docx>=1.2.0  # Word document extraction

# Web Scraping
beautifulsoup4>=4.14.2  # HTML parsing
markdownify>=1.2.0      # HTML to markdown conversion
requests>=2.31.0        # HTTP requests
soupsieve>=2.8          # CSS selector support for BeautifulSoup

# Existing project dependencies (for reference)
# pyyaml>=6.0  # YAML configuration files
# tiktoken>=0.5.0  # Token counting
# openai>=1.0.0  # Embeddings
